Relocate cmake_cp2k.sh and test matcher classes
