+++ tools/docker/Dockerfile.build_hip_rocm_Mi50
@@ -88,7 +88,7 @@ COPY ./cmake ./cmake
 COPY ./CMakeLists.txt .
 
 # Compile CP2K.
-COPY ./tools/docker/scripts/build_cp2k.sh .
+COPY ./tools/docker/scripts/build_cp2k.sh ./tools/docker/scripts/cmake_cp2k.sh ./
 RUN ./build_cp2k.sh toolchain_hip_Mi50 psmp
 
 # Run build test.
