+++ tools/docker/Dockerfile.test_performance_cuda_P100
@@ -83,7 +83,7 @@ COPY ./cmake ./cmake
 COPY ./CMakeLists.txt .
 
 # Compile CP2K.
-COPY ./tools/docker/scripts/build_cp2k.sh .
+COPY ./tools/docker/scripts/build_cp2k.sh ./tools/docker/scripts/cmake_cp2k.sh ./
 RUN ./build_cp2k.sh toolchain_cuda_P100 psmp
 
 # Run performance test for toolchain_cuda_P100.
