+++ tools/docker/scripts/cmake_cp2k.sh
